Move implementations of Resolver, WorkerThread, and ReusableThread to new src files.
Move URI handling in Resolver to separate method.
